requirements.txt
datasets
evaluate
numpy
pillow
requests
torch
transformers
